Fixed a few issues, like removed the host permissions. Moved the api requests to a content script (fiat.js) and made the requests go through api.js
